backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.75.0
anyio==4.11.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
bcrypt==5.0.0
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
cryptography==46.0.3
Deprecated==1.3.1
distro==1.9.0
docstring_parser==0.17.0
ecdsa==0.19.1
elementpath==5.0.4
et_xmlfile==2.0.0
exceptiongroup==1.3.1
fastapi==0.122.0
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-generativeai==0.8.5
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
idna==3.11
jiter==0.12.0
limits==4.2
msal==1.34.0
numpy==2.0.2
openpyxl==3.1.5
packaging==24.2
pandas==2.3.3
passlib==1.7.4
pillow==11.3.0
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
PyJWT==2.10.1
pyparsing==3.2.5
pypdf==6.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.20
pytz==2025.2
reportlab==4.4.5
requests==2.32.5
rsa==4.9.1
sepaxml==2.7.0
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.49.3
text-unidecode==1.3
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.38.0
wrapt==2.0.1
xmlschema==4.2.0
